# Conflicts:
#	src/frontend/devops-repository/src/store/actions/repoGeneric.js
#	src/frontend/locale/repository/en-US.json
#	src/frontend/locale/repository/zh-CN.json

+++ src/backend/common/common-storage/storage-service/src/main/kotlin/com/tencent/bkrepo/common/storage/filesystem/cleanup/CleanupFileVisitor.kt
@@ -37,9 +37,9 @@ import org.slf4j.LoggerFactory
 import java.io.IOException
 import java.nio.file.FileVisitResult
 import java.nio.file.Files
+import java.nio.file.NoSuchFileException
 import java.nio.file.Path
 import java.nio.file.attribute.BasicFileAttributes
-import java.time.Duration
 
 /**
  * 文件清理visitor
@@ -52,6 +52,7 @@ class CleanupFileVisitor(
     private val fileStorage: FileStorage,
     private val fileLocator: FileLocator,
     private val credentials: StorageCredentials,
+    private val fileExpireResolver: FileExpireResolver,
 ) : ArtifactFileVisitor() {
 
     val result = CleanupResult()
@@ -61,13 +62,16 @@ class CleanupFileVisitor(
     @Throws(IOException::class)
     override fun visitFile(filePath: Path, attributes: BasicFileAttributes): FileVisitResult {
         val size = attributes.size()
+        val isTempFile = isTempFile(filePath)
+        var deleted = false
         try {
-            if (isExpired(attributes, expireDuration) && !isNFSTempFile(filePath)) {
-                if (isTempFile(filePath) || existInStorage(filePath)) {
+            if (fileExpireResolver.isExpired(filePath.toFile()) && !isNFSTempFile(filePath)) {
+                if (isTempFile || existInStorage(filePath)) {
                     rateLimiter.acquire()
                     Files.delete(filePath)
                     result.cleanupFile += 1
                     result.cleanupSize += size
+                    deleted = true
                     logger.info("Clean up file[$filePath], size[$size], summary: $result")
                 }
             }
@@ -77,6 +81,11 @@ class CleanupFileVisitor(
         } finally {
             result.totalFile += 1
             result.totalSize += size
+            if(!isTempFile && !deleted) {
+                // 仅统计非temp目录下未被清理的文件
+                result.rootDirNotDeletedFile += 1
+                result.rootDirNotDeletedSize += size
+            }
         }
         return FileVisitResult.CONTINUE
     }
@@ -94,11 +103,14 @@ class CleanupFileVisitor(
         if (dirPath == rootPath || dirPath == tempPath || dirPath == stagingPath) {
             return FileVisitResult.CONTINUE
         }
-        Files.newDirectoryStream(dirPath).use {
-            if (!it.iterator().hasNext()) {
-                Files.delete(dirPath)
-                logger.info("Clean up folder[$dirPath].")
-                result.cleanupFolder += 1
+        // 由于支持删除上传路径，所以这里即使是空目录，也需要判断过期时间。
+        if (fileExpireResolver.isExpired(dirPath.toFile())) {
+            Files.newDirectoryStream(dirPath).use {
+                if (!it.iterator().hasNext()) {
+                    Files.delete(dirPath)
+                    logger.info("Clean up folder[$dirPath].")
+                    result.cleanupFolder += 1
+                }
             }
         }
         result.totalFolder += 1
@@ -109,15 +121,20 @@ class CleanupFileVisitor(
         return expireDuration.seconds > 0
     }
 
-    /**
-     * 判断文件是否过期
-     * 根据上次访问时间和上次修改时间判断
-     */
-    private fun isExpired(attributes: BasicFileAttributes, expireDuration: Duration): Boolean {
-        val lastAccessTime = attributes.lastAccessTime().toMillis()
-        val lastModifiedTime = attributes.lastModifiedTime().toMillis()
-        val expiredTime = System.currentTimeMillis() - expireDuration.toMillis()
-        return lastAccessTime < expiredTime && lastModifiedTime < expiredTime
+    override fun visitFileFailed(file: Path, exc: IOException?): FileVisitResult {
+        if (exc is NoSuchFileException) {
+            // 目录或者文件已经由其他进程删除。
+            logger.info("File [$file] already delete.")
+        } else {
+            logger.error("Clean up file [$file] error.", exc)
+            result.errorCount++
+            if (Files.isRegularFile(file)) {
+                result.totalFile++
+            } else {
+                result.totalFolder++
+            }
+        }
+        return FileVisitResult.CONTINUE
     }
 
     /**
